docs: README.md 회고 && refactoring complete
