+++ .env
@@ -1,2 +1,2 @@
-REACT_APP_GITHUB_ACCESS_TOKEN = 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
+REACT_APP_GITHUB_ACCESS_TOKEN = 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
 REACT_APP_GITHUB_API_URL = https://api.github.com
