+++ src/api/IssueDetailContext.tsx
@@ -1,6 +1,7 @@
 import { createContext } from 'react';
 import { fetchIssueType, useFetchDetail } from '../hooks/useFetchDetail';
 
+// 컨텍스트 생성
 export const IssueDetailContext = createContext({
   issueDetail: {
     avatar_url: undefined,
@@ -16,6 +17,7 @@ export const IssueDetailContext = createContext({
   fetchIssueDetail: (id: string) => {},
 });
 
+// 이슈의 세부 내용 상태를 관리
 export const IssueDetailProvider = ({ children }: any) => {
   const { issueDetail, loading, error, fetchIssueDetail } = useFetchDetail();
