--- src/hooks/useIssue.ts
@@ -1,8 +1,0 @@
-import { useContext } from 'react';
-import { IssueContext } from '../api/IssueContext';
-
-export const useIssues = () => {
-  const { issues, loading, error } = useContext(IssueContext);
-
-  return { issues, loading, error };
-};
